refactor(registry): move demos to shadcn examples/<name>-demo convention
Relocate the 28 component demos from co-located
`registry/new-york/<name>/<name>.demo.tsx` to a single
`registry/new-york/examples/<name>-demo.tsx` folder, matching shadcn's
registry layout (ui/ + blocks/ + examples/).

- git mv each demo; remove the now-empty per-component dirs.
- registry-meta.ts: update the 28 demo imports.
- [component] route: read the demo source from the new examples path; the
  Usage code tab label is now `<name>-demo.tsx`.

tsc --noEmit and next build pass; all examples map to a registry entry.
