+++ registry/new-york/examples/animated-number-demo.tsx
