+++ registry/new-york/examples/announcement-bar-demo.tsx
