+++ registry/new-york/examples/avatar-stack-demo.tsx
