+++ registry/new-york/examples/color-picker-demo.tsx
