+++ registry/new-york/examples/copy-button-demo.tsx
