+++ registry/new-york/examples/currency-input-demo.tsx
