+++ registry/new-york/examples/empty-state-demo.tsx
