+++ registry/new-york/examples/file-dropzone-demo.tsx
