+++ registry/new-york/examples/kbd-demo.tsx
