+++ registry/new-york/examples/marquee-demo.tsx
