+++ registry/new-york/examples/month-picker-demo.tsx
