+++ registry/new-york/examples/number-range-demo.tsx
