+++ registry/new-york/examples/password-input-demo.tsx
