+++ registry/new-york/examples/phone-input-demo.tsx
