+++ registry/new-york/examples/rating-demo.tsx
