+++ registry/new-york/examples/scroll-progress-demo.tsx
