+++ registry/new-york/examples/spinner-demo.tsx
