+++ registry/new-york/examples/stat-card-demo.tsx
