+++ registry/new-york/examples/stepper-demo.tsx
