+++ registry/new-york/examples/year-picker-demo.tsx
