+++ registry/new-york/registry-meta.ts
@@ -1,33 +1,33 @@
 import * as React from "react"
 
-import AnimatedNumberDemo from "@/registry/new-york/animated-number/animated-number.demo"
-import AnnouncementBarDemo from "@/registry/new-york/announcement-bar/announcement-bar.demo"
-import AvatarStackDemo from "@/registry/new-york/avatar-stack/avatar-stack.demo"
-import CalloutDemo from "@/registry/new-york/callout/callout.demo"
-import ColorPickerDemo from "@/registry/new-york/color-picker/color-picker.demo"
-import ComboboxDemo from "@/registry/new-york/combobox/combobox.demo"
-import CopyButtonDemo from "@/registry/new-york/copy-button/copy-button.demo"
-import CurrencyInputDemo from "@/registry/new-york/currency-input/currency-input.demo"
-import EmptyStateDemo from "@/registry/new-york/empty-state/empty-state.demo"
-import FileDropzoneDemo from "@/registry/new-york/file-dropzone/file-dropzone.demo"
-import KbdDemo from "@/registry/new-york/kbd/kbd.demo"
-import LazySelectDemo from "@/registry/new-york/lazy-select/lazy-select.demo"
-import MarqueeDemo from "@/registry/new-york/marquee/marquee.demo"
-import MonthPickerDemo from "@/registry/new-york/month-picker/month-picker.demo"
-import MultiSelectDemo from "@/registry/new-york/multi-select/multi-select.demo"
-import NumberRangeDemo from "@/registry/new-york/number-range/number-range.demo"
-import PasswordInputDemo from "@/registry/new-york/password-input/password-input.demo"
-import PhoneInputDemo from "@/registry/new-york/phone-input/phone-input.demo"
-import RatingDemo from "@/registry/new-york/rating/rating.demo"
-import ScrollProgressDemo from "@/registry/new-york/scroll-progress/scroll-progress.demo"
-import SpinnerDemo from "@/registry/new-york/spinner/spinner.demo"
-import StatCardDemo from "@/registry/new-york/stat-card/stat-card.demo"
-import StepperDemo from "@/registry/new-york/stepper/stepper.demo"
-import TagInputDemo from "@/registry/new-york/tag-input/tag-input.demo"
-import TimelineDemo from "@/registry/new-york/timeline/timeline.demo"
-import TimePickerDemo from "@/registry/new-york/time-picker/time-picker.demo"
-import TreeViewDemo from "@/registry/new-york/tree-view/tree-view.demo"
-import YearPickerDemo from "@/registry/new-york/year-picker/year-picker.demo"
+import AnimatedNumberDemo from "@/registry/new-york/examples/animated-number-demo"
+import AnnouncementBarDemo from "@/registry/new-york/examples/announcement-bar-demo"
+import AvatarStackDemo from "@/registry/new-york/examples/avatar-stack-demo"
+import CalloutDemo from "@/registry/new-york/examples/callout-demo"
+import ColorPickerDemo from "@/registry/new-york/examples/color-picker-demo"
+import ComboboxDemo from "@/registry/new-york/examples/combobox-demo"
+import CopyButtonDemo from "@/registry/new-york/examples/copy-button-demo"
+import CurrencyInputDemo from "@/registry/new-york/examples/currency-input-demo"
+import EmptyStateDemo from "@/registry/new-york/examples/empty-state-demo"
+import FileDropzoneDemo from "@/registry/new-york/examples/file-dropzone-demo"
+import KbdDemo from "@/registry/new-york/examples/kbd-demo"
+import LazySelectDemo from "@/registry/new-york/examples/lazy-select-demo"
+import MarqueeDemo from "@/registry/new-york/examples/marquee-demo"
+import MonthPickerDemo from "@/registry/new-york/examples/month-picker-demo"
+import MultiSelectDemo from "@/registry/new-york/examples/multi-select-demo"
+import NumberRangeDemo from "@/registry/new-york/examples/number-range-demo"
+import PasswordInputDemo from "@/registry/new-york/examples/password-input-demo"
+import PhoneInputDemo from "@/registry/new-york/examples/phone-input-demo"
+import RatingDemo from "@/registry/new-york/examples/rating-demo"
+import ScrollProgressDemo from "@/registry/new-york/examples/scroll-progress-demo"
+import SpinnerDemo from "@/registry/new-york/examples/spinner-demo"
+import StatCardDemo from "@/registry/new-york/examples/stat-card-demo"
+import StepperDemo from "@/registry/new-york/examples/stepper-demo"
+import TagInputDemo from "@/registry/new-york/examples/tag-input-demo"
+import TimelineDemo from "@/registry/new-york/examples/timeline-demo"
+import TimePickerDemo from "@/registry/new-york/examples/time-picker-demo"
+import TreeViewDemo from "@/registry/new-york/examples/tree-view-demo"
+import YearPickerDemo from "@/registry/new-york/examples/year-picker-demo"
 
 import AppShell01 from "@/registry/new-york/blocks/app-shell-01/app-shell-01"
 import AppShell02 from "@/registry/new-york/blocks/app-shell-02/app-shell-02"
